requirements.txt
apache-airflow
teradatasql
